+++ Tests/PSRemoteOperations.Tests.ps1
@@ -7,11 +7,12 @@ Param()
 
 $moduleName = (($MyInvocation.mycommand).name -split "\.")[0]
 $ModuleManifestName = "$modulename.psd1"
-$ModuleManifestPath = "$PSScriptRoot\..\..\$ModuleManifestName"
+$ModuleManifestPath = "..\$ModuleManifestName"
 
 If (Get-Module $modulename) {
     Remove-module $moduleName
 }
+Write-Host "Importing $modulemanifestPath" -ForegroundColor yellow
 import-module $ModuleManifestPath -Force
 
 Describe $ModuleName {
@@ -21,8 +22,8 @@ Describe $ModuleName {
         It 'Passes Test-ModuleManifest' {
             $myModule | Should Not BeNullOrEmpty
         }
-        It "Should NOT have a root module" {
-            $myModule.RootModule | Should BeNullOrEmpty
+        It "Should have a root module" {
+            $myModule.RootModule | Should Not BeNullOrEmpty
         }
         It "Contains exported commands" {
             $myModule.ExportedCommands | Should Not BeNullOrEmpty
@@ -84,31 +85,31 @@ Describe $ModuleName {
     }
     Context Structure {
         It "Should have a Docs folder" {
-            Get-Item $PSScriptRoot\..\..\docs | Should Be $True
+            Get-Item $PSScriptRoot\..\docs | Should Be $True
         }
         foreach ($cmd in $myModule.ExportedFunctions.keys) {
             It "Should have a markdown help file for $cmd" {
-                "$PSScriptRoot\..\..\docs\$cmd.md" | Should Exist
+                "$PSScriptRoot\..\docs\$cmd.md" | Should Exist
             }
         }
         It "Should have an external help file" {
-            "$PSScriptRoot\..\..\en-us\*.xml" | Should Exist
-            "$PSScriptRoot\..\..\en-us\*.txt" | Should Exist
+            "$PSScriptRoot\..\en-us\*.xml" | Should Exist
+            "$PSScriptRoot\..\en-us\*.txt" | Should Exist
         }
 
         It "Should have an about file" {
-            "$PSScriptRoot\..\..\docs\about_$ModuleName.md"| Should Exist
+            "$PSScriptRoot\..\docs\about_$ModuleName.md"| Should Exist
         }
         It "Should have a license file" {
-            "$PSScriptRoot\..\..\license.*" | Should Exist
+            "$PSScriptRoot\..\license.*" | Should Exist
         }
 
         It "Should have a changelog file" {
-            "$PSScriptRoot\..\..\changelog*" | Should Exist
+            "$PSScriptRoot\..\changelog*" | Should Exist
         }
 
         It "Should have a README.md file" {
-            "$PSScriptRoot\..\..\README.md" | Should Exist
+            "$PSScriptRoot\..\README.md" | Should Exist
         }
     }
 } -Tag module
@@ -126,16 +127,16 @@ InModuleScope PSRemoteOperations {
             ($params["scriptpath"].attributes).where( {$_.TypeId -match 'parameter'}).Mandatory | Should Be $True
         }
         It "Should create a psd1 file" {
-            {New-PSRemoteOperation -Computername Foo -Path TestDrive:\ -Scriptblock {1}} | Should Not Throw
-            Test-Path TestDrive:\foo*.psd1 | Should be $True
+            {New-PSRemoteOperation -Computername Foo -Path TESTDRIVE:\ -Scriptblock {1}} | Should Not Throw
+            Test-Path TESTDRIVE:\foo*.psd1 | Should be $True
 
             #Should have a filename using the pattern computername_guid.psd1"
             $guidrx = "[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}"
             $template = "foo_$guidrx.psd1"
-            (Get-Item Testdrive:\*.psd1).name -match $Template | Should be $True
+            (Get-Item TESTDRIVE:\*.psd1).name -match $Template | Should be $True
 
             #get the file contents to use in the next set of tests
-            $script:in = Import-PowerShellDataFile Testdrive:\foo_*.psd1
+            $script:in = Import-PowerShellDataFile TESTDRIVE:\foo_*.psd1
         }
 
         It "Should create a file with a Computername value of Foo" {
@@ -150,45 +151,70 @@ InModuleScope PSRemoteOperations {
         }
 
         It "Should accept multiple computernames" {
-            $files = New-PSRemoteOperation -Computername Foo1, Foo2, Foo3 -Path TestDrive:\ -Scriptblock {Get-Volume C:} -Passthru
+            $files = New-PSRemoteOperation -Computername Foo1, Foo2, Foo3 -Path TESTDRIVE:\ -Scriptblock {Get-Volume C:} -Passthru
             $files.count | Should Be 3
         }
     } -tag command
 
     Describe Invoke-PSRemoteOperation {
-        New-PSRemoteOperation -Computername $env:computername -Path TestDrive: -Scriptblock {Get-Service bits}
-        New-Item -path TestDrive:\ -Name Archive -ItemType directory
+        mock _psInvoke {}
+        New-PSRemoteOperation -Computername FOO -Path TESTDRIVE: -Scriptblock {Get-Service bits} -passthru
+        <#
+        @{
+            CreatedOn = 'Prospero'
+            CreatedBy = 'prospero\jeff'
+            CreatedAt = '11/13/2020 19:13:25 UTC'
+            Computername = 'FOO'
+            PSVersion = 5.1
+            Scriptblock = 'Get-Service bits'
+            }
+        #>
+        New-Item -path TESTDRIVE:\ -Name Archive -ItemType directory
 
         Context Input {
             $params = (Get-Command Invoke-PSRemoteOperation).parameters
             It "Should have a mandatory Path parameter ending in .psd1" {
-                ($params["Path"].attributes).where( {$_.TypeId -match 'parameter'}).Mandatory | Should Be $True
-                ($params["Path"].attributes).where( {$_.TypeId -match 'pattern'}).regexPattern | Should Be '\.psd1$'
+                ($params["Path"].attributes).where({$_.TypeId -match 'parameter'}).Mandatory | Should Be $True
+                ($params["Path"].attributes).where({$_.TypeId -match 'pattern'}).regexPattern | Should Be '\.psd1$'
             }
         }
         Context Process {
-            $file = (Get-childitem testdrive:\*.psd1).fullname
 
-            Mock 'Get-Date' {[datetime]::new(2019, 1, 1)}
-
-            #I'd like to mock this but there doesn't appear to be a way
-            #Mock New-PSSession {}
+            $file = (Get-ChildItem TESTDRIVE:\*.psd1).fullname
 
             Mock Convert-Path { $file }
+            Mock Get-Service {$True}
+            #write-host "Mock invoking $file" -ForegroundColor cyan
+            Invoke-PSRemoteOperation -Path $file -ArchivePath TESTDRIVE:\archive #-Verbose
 
-            Invoke-PSRemoteOperation -Path $file -ArchivePath testdrive:\archive #-Verbose
+            $mockresult = @"
+@{
+CreatedOn = 'Prospero'
+CreatedBy = 'prospero\jeff'
+CreatedAt = '11/13/2020 19:00:00 UTC'
+Computername = 'FOO'
+PSVersion = '5.1'
+Scriptblock = 'Get-Service bits'
+Completed = 'True'
+Error = ""
+Date = '11/13/2020 19:00:00 UTC'
+}
+"@
+            $mockresult | Out-File TESTDRIVE:\Archive\foo.psd1
 
             It "Should call Convert-Path" {
                 Assert-MockCalled Convert-Path
             }
 
-            It "Should call Get-Date" {
-                Assert-MockCalled Get-Date
+            It "Should call _psInvoke" {
+                Assert-MockCalled _psInvoke
             }
+
             #save result for the next set of tests
-            $script:result = Get-ChildItem TestDrive:\archive\*.psd1
-            $script:rcontent = get-item $script:result | get-content
-            $script:rdata = Import-powershelldatafile $script:result
+            $script:result = Get-ChildItem TESTDRIVE:\archive\*.psd1
+            $script:rcontent = Get-Item $script:result | Get-Content
+            #write-host $script:rcontent -ForegroundColor cyan
+            $script:rdata = Import-PowerShellDataFile $script:result
         }
 
         Context Output {
@@ -200,21 +226,24 @@ InModuleScope PSRemoteOperations {
                 $script:rdata.Completed | Should be 'True'
             }
             It "Should show the local computername" {
-                $script:rdata.Computername | Should be $env:computername
+                $script:rdata.Computername | Should be "FOO"
             }
             It "Should show the scriptblock" {
                 $script:rdata.Scriptblock.trim() | Should be "Get-Service bits"
             }
             It "Should show a date of 01/01/2019" {
-                $script:rdata.Date -match "01\/01\/2019\s(\d{2}:){2}00\sUTC" | Should be $True
+                $script:rdata.Date -match "11\/13\/2020\s(\d{2}:){2}00\sUTC" | Should be $True
+            }
+            It "Should have run under PowerShell 5.1" {
+                $script:rdata.PSVersion | Should be "5.1"
             }
         }
     } -tag command
 
     Describe Get-PSRemoteOperationResult {
         #copy the results from an earlier test to avoid duplication
-        New-Item -path TestDrive: -Name Archive -ItemType directory
-        $fake = Join-Path -Path TestDrive:\Archive -ChildPath "$($env:computername)_$(New-Guid).psd1"
+        New-Item -path TESTDRIVE: -Name Archive -ItemType directory
+        $fake = Join-Path -Path TESTDRIVE:\Archive -ChildPath "$($env:computername)_$(New-Guid).psd1"
 
         $script:rcontent | Out-File -FilePath $fake
 
@@ -236,7 +265,7 @@ InModuleScope PSRemoteOperations {
     } -tag command
 
     Describe Register-PSRemoteOperationWatcher {
-        New-Item -path TestDrive: -Name Archive -ItemType directory
+        New-Item -path TESTDRIVE: -Name Archive -ItemType directory
 
         Mock New-JobTrigger { [Microsoft.PowerShell.ScheduledJob.ScheduledJobTrigger]::new()}
         Mock Register-ScheduledJob {}
@@ -246,8 +275,8 @@ InModuleScope PSRemoteOperations {
         $testparams = @{
             Credential  = $cred
             Name        = 'TestWatch'
-            Path        = 'Testdrive:\'
-            ArchivePath = 'Testdrive:\Archive'
+            Path        = 'TESTDRIVE:\'
+            ArchivePath = 'TESTDRIVE:\Archive'
         }
         Register-PSRemoteOperationWatcher @testparams