+++ autocompleters.ps1
@@ -2,7 +2,7 @@
 Register-ArgumentCompleter -CommandName Get-PSRemoteOperationResult, New-PSRemoteOperation -ParameterName Computername -ScriptBlock {
     param($commandName, $parameterName, $wordToComplete, $commandAst, $fakeBoundParameter)
 
-    $names = (Get-ChildItem -Path $PSRemoteOpArchive | Split-Path -leaf).foreach( {$_.split("_")[0].toUpper() }) | Get-Unique
+    $names = (Get-ChildItem -Path $PSRemoteOpArchive | Split-Path -leaf).foreach( { $_.split("_")[0].toUpper() }) | Get-Unique
 
     if ($wordToComplete) {
         $fill = $names | Where-Object { $_ -match "$wordToComplete" }
@@ -18,7 +18,7 @@ Register-ArgumentCompleter -CommandName Get-PSRemoteOperationResult, New-PSRemot
 Register-ArgumentCompleter -CommandName Wait-PSRemoteOperation, Get-PSRemoteOperation -ParameterName Computername -ScriptBlock {
     param($commandName, $parameterName, $wordToComplete, $commandAst, $fakeBoundParameter)
 
-    $names = (Get-ChildItem -Path $PSRemoteOpPath -file | Split-Path -leaf).foreach( {$_.split("_")[0].toUpper() }) | Get-Unique
+    $names = (Get-ChildItem -Path $PSRemoteOpPath -file | Split-Path -leaf).foreach( { $_.split("_")[0].toUpper() }) | Get-Unique
 
     if ($wordToComplete) {
         $fill = $names | Where-Object { $_ -match "$wordToComplete" }
@@ -36,8 +36,8 @@ Register-ArgumentCompleter -CommandName New-PSRemoteOperation -ParameterName To 
 
     (Get-ChildItem -Path Cert:\CurrentUser\my -DocumentEncryptionCert).Subject |
         ForEach-Object {
-        [System.Management.Automation.CompletionResult]::new($_, $_, 'ParameterValue', $_)
-    }
+            [System.Management.Automation.CompletionResult]::new($_, $_, 'ParameterValue', $_)
+        }
 }
 
 <#
@@ -54,11 +54,11 @@ else {
 No PSRemoteOpPath settings file was found to import. The module will be easier to use
 if you set global variables for `$PSRemoteOpPath and `$PSRemoteOpArchive. Run the
 commmand Register-PSRemoteOpPath. If you've recently updated the module through the
-PowerShell Gallery, you'll need to re-register the paths.
+PowerShell Gallery, you might need to re-register the paths.
 
 If you prefer to set the values in your profile script you can ignore this
 warning.
-"
+
 "@
     Write-Warning $msg
 }