+++ functions/functions.ps1
@@ -41,7 +41,12 @@ Function New-PSRemoteOperation {
         [Parameter(HelpMessage = "The folder where the remote operation file will be created.")]
         [string]$Path = $PSRemoteOpPath,
 
-        [switch]$Passthru
+        [switch]$Passthru,
+
+        [Parameter(HelpMessage = "Specify which version of PowerShell to use for the remote operation.")]
+        [ValidateSet("Desktop", "Core")]
+        [string]$PSVersion = "Desktop"
+
     )
     DynamicParam {
         if (Get-Command Protect-CmsMessage -ea silentlycontinue) {
@@ -68,18 +73,22 @@ Function New-PSRemoteOperation {
         foreach ($Computer in $Computername) {
             Write-Verbose "Creating a remote operations file for $($computer.toUpper())"
             #define a here string for the psd1 content
+            $Who = "$([System.Environment]::UserDomainName)\$([system.environment]::UserName)"
+
             $out = @"
 @{
-CreatedOn = '$(hostname)'
-CreatedBy = '$(whoami)'
-CreatedAt = '$((Get-Date).toUniversalTime()) UTC'
+CreatedOn = '$([System.Environment]::Machinename)'
+CreatedBy = '$($Who)'
+CreatedAt = '$((Get-Date).toUniversalTime())'
 Computername = '$($Computer.ToUpper())'
+PSVersion = $PSVersion
+Status = 'Pending'
 
 "@
 
             if ($ArgumentList) {
 
-                $opArgs = Convert-HashTableToCode $ArgumentList
+                $opArgs = Convert-HashtableToCode $ArgumentList
                 $out += "ArgumentList = $opargs"
                 $out += "`n"
             }
@@ -102,8 +111,8 @@ Computername = '$($Computer.ToUpper())'
             $out | Write-Verbose
 
             #make the filename all lower case
-            $fname = "$($Computer.ToUpper())_$(New-GUID).psd1"
-            $outFile = Join-path -Path $Path -ChildPath $fname #.toLower()
+            $fname = "$($Computer.ToUpper())_$(New-Guid).psd1"
+            $outFile = Join-Path -Path $Path -ChildPath $fname #.toLower()
 
             Write-Verbose "Creating datafile $outfile"
             if ($PSCmdlet.ShouldProcess($outFile, "Creating PSRemote Operations File")) {
@@ -113,10 +122,10 @@ Computername = '$($Computer.ToUpper())'
                     Protect-CmsMessage -To $PSBoundParameters.Item("to") -Content $out -OutFile $outFile
                 }
                 else {
-                    $out | Out-File -FilePath $outFile -force -Encoding ascii
+                    $out | Out-File -FilePath $outFile -Force -Encoding ascii
                 }
                 if ($Passthru) {
-                    Get-Item -path $outFile
+                    Get-Item -Path $outFile
                 }
             } #if should process
         } #foreach computer
@@ -155,6 +164,162 @@ Function Invoke-PSRemoteOperation {
     } #begin
 
     Process {
+        #this code was updated 13 Nov 2020 to allow running different PowerShell version (Issue #12)
+
+        #convert paths to plain filesystem paths
+        $cPath = Convert-Path $path
+        Write-Verbose "Processing $cPath"
+
+        $parent = Split-Path -Path $cPath -Parent
+        Write-Verbose "Comparing $parent to $ArchivePath"
+        #The archive path and path for data file must be different
+        if ($parent -eq $ArchivePath) {
+            Write-Warning "The archive path must be different from the path to the psd1 file."
+            #bail out
+            Return
+        }
+
+        #test if the file contains a CMS message
+        Try {
+            Write-Verbose "Testing if a CMS Message"
+            $cms = Get-CmsMessage -Path $cPath -ErrorAction stop
+            $to = $cms.Recipients.issuerName
+            Write-Verbose "Detected CMS recipient $to"
+            $raw = (Unprotect-CmsMessage -Path $cPath).split("`n")
+            $in = $raw | Out-String | Convert-HashtableString
+        }
+        Catch {
+            Write-Verbose "Not a CMS Message or this is a non-Windows platform. $($_.exception.message)"
+            Write-Verbose "Import the data file to create a settings hashtable"
+            $in = Import-PowerShellDataFile -Path $cPath
+            $raw = Get-Content -Path $cpath
+            Write-Verbose ($raw | Out-String)
+            $To = $False
+        }
+
+        Write-Verbose "Using content:"
+        $in | Out-String | Write-Verbose
+
+        if ($PSCmdlet.ShouldProcess($cPath)) {
+
+            <#
+            This is probably a bit overly complicated, but it works and supports
+            running the operation in different PowerShell versions, so I'm
+            happy to let it be.
+            #>
+            $sb = {
+                param($in, $cpath, $ArchivePath, $raw, $to)
+
+                #uncomment for troubleshooting and development
+                # $VerbosePreference = "Continue"
+                # Start-Transcript c:\work\sb.log
+                #set hashtable values to correct type
+                if ($in.Scriptblock) {
+                    Write-Verbose "Creating scriptblock"
+                    $action = [scriptblock]::Create($in.Scriptblock)
+                }
+                else {
+                    Write-Verbose "Getting script contents from $($in.FilePath)"
+                    $action = Get-Content -Path $in.FilePath | Out-String
+                }
+
+                if ($in.ArgumentList) {
+                    #arguments must be entered as a hashtable
+                    Write-Verbose "Adding Parameters"
+                    $actionParams = $in.ArgumentList
+
+                    #convert True/False to switches.
+                    $bool2switch = $actionparams.keys | Where-Object { $actionparams[$_] -match 'true|false' }
+                    if ($bool2switch) {
+                        foreach ($item in $bool2Switch) {
+                            Write-Host "Treating $item as a switch"
+                            $txt = $actionParams[$item]
+                            Write-Host "converting $txt"
+                            if ($txt -eq 'true') {
+                                $asSwitch = $True -as [switch]
+                            }
+                            else {
+                                $asSwitch = $False -as [switch]
+                            }
+                            $actionParams[$item] = $asSwitch
+                        }
+                    }
+                    Write-Verbose ($actionParams | Out-String)
+                }
+
+                $psrunspace = [powershell]::Create()
+
+                if ($in.Initialization) {
+                    Write-Verbose "Adding initialization"
+                    $init = [scriptblock]::Create($in.Initialization)
+                    [void]$psrunspace.AddScript($init)
+                }
+
+                Write-Verbose "Adding action"
+                [void]$psrunspace.Addscript($action)
+
+                if ($actionParams) {
+                    Write-Verbose "Adding parameters"
+                    [void]$psrunspace.AddParameters($actionParams)
+                }
+
+                Write-Verbose ($psrunspace.Commands.commands | Out-String)
+                $psrunspace.invoke()
+                if ($psrunspace.HadErrors) {
+                    $completed = $False
+                }
+                else {
+                    $completed = $True
+                }
+
+                $errormsg = """$($psrunspace.Streams.Error.exception.Message)"""
+                $psrunspace.dispose()
+
+                #create a results file
+                Write-Verbose "Result Data"
+                $resultdata = @"
+@{
+
+"@
+                #append the result data to the data file.
+                ($Raw | Select-Object -Skip 1 | Select-Object -SkipLast 1 ).Foreach( { $resultData += "$_`n" })
+
+                $resultData += "Completed = '$completed'`n"
+                #replace any variables in the errormessage with escaped literals
+                $errormsg = $errormsg.Replace('$', '`$')
+                $resultData += "Error = $errormsg`n"
+                $resultdata += "Date = '$((Get-Date).toUniversalTime()) UTC'`n"
+
+                $resultData += "}"
+
+                $resultdata | Out-String | Write-Verbose
+
+                $filename = Split-Path -Path $cPath -Leaf
+                $resultFile = Join-Path -Path $ArchivePath -ChildPath $filename
+
+                Write-Verbose "Creating results file $resultFile"
+                if ($to) {
+                    Write-Verbose "Create a CMS archive file to $To"
+                    Protect-CmsMessage -Content $resultdata -To $to -OutFile $resultFile
+                }
+                else {
+                    $resultdata | Out-File -FilePath $resultFile -Encoding ascii
+                }
+                #delete
+                Write-Verbose "Removing operation file $cPath"
+                Remove-Item -Path $cPath -Force
+            } #cmd scriptblock
+
+            #invoke a private function to actually run the commands.
+            #using a function so it can be mocked for Pester tests
+            #powershell.exe -noprofile -command $sb -args $in, $cpath, $ArchivePath, $raw, $to
+            _psInvoke -scriptblock $sb -parameters @($in, $cpath, $ArchivePath, $raw, $to)
+
+        } #should process
+    }#process
+
+    #old code
+    <#     Process {
         #convert paths to plain filesystem paths
         $cPath = Convert-Path $path
         Write-Verbose "Processing $cPath"
@@ -174,14 +339,14 @@ Function Invoke-PSRemoteOperation {
             $cms = Get-CmsMessage -Path $cPath -ErrorAction stop
             $to = $cms.Recipients.issuerName
             $raw = (Unprotect-CmsMessage -path $cPath).split("`n")
-            $in = $raw | Out-string | Convert-HashtableString
+            $in = $raw | Out-String | Convert-HashtableString
         }
         Catch {
             Write-Verbose "Not a CMS Message or this is a non-Windows platform. $($_.exception.message)"
             Write-Verbose "Import the data file to create a settings hashtable"
             $in = Import-PowerShellDataFile -Path $cPath
             $raw = Get-Content -Path $cpath
-            Write-Verbose ($raw | Out-string)
+            Write-Verbose ($raw | Out-String)
             $To = $False
         }
 
@@ -206,7 +371,7 @@ Function Invoke-PSRemoteOperation {
 
         if ($PSCmdlet.ShouldProcess($cPath)) {
 
-            #TODO - this should be turned into private functions so it can be tested with Pester
+
             $psrunspace = [powershell]::Create()
             if ($in.Initialization) {
                 Write-Verbose "Adding initialization"
@@ -239,7 +404,7 @@ Function Invoke-PSRemoteOperation {
 
 "@
             #append the result data to the data file.
-            ($Raw | Select-Object -skip 1 | Select-Object -SkipLast 1 ).Foreach( { $resultData += "$_`n" })
+            ($Raw | Select-Object -skip 1 | Select-Object -SkipLast 1 ).Foreach({ $resultData += "$_`n" })
 
             $resultData += "Completed = '$completed'`n"
             #replace any variables in the errormessage with escaped literals
@@ -265,10 +430,8 @@ Function Invoke-PSRemoteOperation {
             #delete
             Write-Verbose "Removing operation file $cPath"
             Remove-Item -Path $cPath -Force
-
-            # } #finally
         } #should process
-    }#process
+    }#process #>
 
     End {
         Write-Verbose "Ending $($MyInvocation.MyCommand)"
@@ -297,7 +460,7 @@ Function Get-PSRemoteOperationResult {
             ParameterSetName = "result"
         )]
         [Parameter(ParameterSetName = "raw")]
-        [ValidateScript( { Test-Path -path $_ })]
+        [ValidateScript( { Test-Path -Path $_ })]
         [Alias("path")]
         [string]$ArchivePath = $PSRemoteOpArchive,
 
@@ -324,7 +487,7 @@ Function Get-PSRemoteOperationResult {
     }
 
     Write-Verbose "Filtering for $filter"
-    $data = Get-ChildItem -Path $ArchivePath -filter $filter | Sort-Object -Property LastWriteTime -Descending
+    $data = Get-ChildItem -Path $ArchivePath -Filter $filter | Sort-Object -Property LastWriteTime -Descending
 
     if ($Newest -gt 0) {
         Write-Verbose "Getting newest $newest results"
@@ -351,7 +514,7 @@ Function Get-PSRemoteOperationResult {
         }
         else {
             $chash.Add("Path", $file.fullname)
-            $obj = New-Object -typename psobject -Property $chash
+            $obj = New-Object -TypeName psobject -Property $chash
             $obj.psobject.typenames.insert(0, "RemoteOpResult")
             $obj
         }
@@ -429,7 +592,7 @@ Function Wait-PSRemoteOperation {
                 Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Timeout exceeded"
                 Break
             }
-        } while (Test-Path -path $Target )
+        } while (Test-Path -Path $Target )
         Write-Verbose "[$((Get-Date).TimeofDay) PROCESS] Total waiting time $timer seconds."
     } #process
 
@@ -457,7 +620,7 @@ Function Get-PSRemoteOperation {
             Position = 1,
             HelpMessage = "Enter the path to the operations folder."
         )]
-        [ValidateScript( { Test-Path -path $_ })]
+        [ValidateScript( { Test-Path -Path $_ })]
         [string]$Path = $PSRemoteOpPath
     )
 
@@ -472,7 +635,7 @@ Function Get-PSRemoteOperation {
     }
 
     Write-Verbose "Filtering for $filter"
-    $data = Get-ChildItem -Path $Path -filter $filter | Sort-Object -Property LastWriteTime -Descending
+    $data = Get-ChildItem -Path $Path -Filter $filter | Sort-Object -Property LastWriteTime -Descending
 
     foreach ($file in $data) {
         Write-Verbose "Processing $($file.fullname)"
@@ -492,10 +655,25 @@ Function Get-PSRemoteOperation {
         }
 
         if ($chash) {
-            $chash.Add("Path", $file.fullname)
-            $obj = New-Object -typename psobject -Property $chash
+            #convert hashtable into a custom object
+            [pscustomobject]@{
+                PSTypename   = "PSRemoteOp"
+                PSVersion    = $chash.PSVersion
+                CreatedOn    = $chash.CreatedOn
+                StartTime    = $null
+                Scriptblock  = $chash.Scriptblock
+                Runtime      = $null
+                EndTime      = $null
+                Computername = $chash.computername
+                CreatedBy    = $chash.CreatedBy
+                CreatedAt    = $chash.CreatedAt
+                Status       = $chash.status
+
+            }
+            <# $chash.Add("Path", $file.fullname)
+            $obj = New-Object -TypeName psobject -Property $chash
             $obj.psobject.typenames.insert(0, "RemoteOp")
-            $obj
+            $obj #>
         }
         #clear the variable so it doesn't accidently get re-used
         Remove-Variable chash -ErrorAction SilentlyContinue
@@ -508,29 +686,29 @@ Function Register-PSRemoteOpPath {
     [cmdletbinding(SupportsShouldProcess)]
     Param(
         [Parameter(Mandatory, HelpMessage = "Enter a filesystem path for the Remote Operations path")]
-        [ValidateScript({ (Test-Path $_ )-AND ((Get-Item $_).psprovider.name -eq "filesystem")})]
+        [ValidateScript( { (Test-Path $_ ) -AND ((Get-Item $_).psprovider.name -eq "filesystem") })]
         [string]$PSRemoteOpPath,
 
         [Parameter(Mandatory, HelpMessage = "Enter a filesystem path for the Remote Operations archive path")]
-        [ValidateScript( { (Test-Path $_) -AND ((Get-Item $_).psprovider.name -eq "filesystem")})]
+        [ValidateScript( { (Test-Path $_) -AND ((Get-Item $_).psprovider.name -eq "filesystem") })]
         [string]$PSRemoteOpArchive
     )
 
     Write-Verbose "Starting $($MyInvocation.MyCommand)"
 
-    $json = Join-Path -path $PSScriptRoot -ChildPath psremoteoppath.json
+    $json = Join-Path -Path $PSScriptRoot -ChildPath psremoteoppath.json
 
     $data = [pscustomobject]@{
         PSRemoteOpPath    = $PSRemoteOpPath
         PSRemoteOpArchive = $PSRemoteOpArchive
-        Updated           = (Get-Date -format f)
+        Updated           = (Get-Date -Format f)
     }
     Write-Verbose "Registering this data"
     $data | Out-String | Write-Verbose
     Write-Verbose "to $json"
 
     if ($PSCmdlet.ShouldProcess($json)) {
-        $data | ConvertTo-Json | Out-File -filepath $json
+        $data | ConvertTo-Json | Out-File -FilePath $json
         #import the data
         Import-PSRemoteOpPath
     } #end whatif
@@ -543,14 +721,14 @@ Function Import-PSRemoteOpPath {
 
     Param(
         [Parameter(HelpMessage = "Enter the path to the remote op path json file.")]
-        [ValidateScript({Test-Path $_})]
-        [string]$Path = (Join-Path -path $PSScriptRoot -ChildPath psremoteoppath.json)
+        [ValidateScript( { Test-Path $_ })]
+        [string]$Path = (Join-Path -Path $PSScriptRoot -ChildPath psremoteoppath.json)
     )
 
     Write-Verbose "Ending $($MyInvocation.MyCommand)"
     Write-Verbose "Importing settings from $path"
 
-    $in = Get-Content -path $json | ConvertFrom-Json
+    $in = Get-Content -Path $json | ConvertFrom-Json
 
     $in | Out-String | Write-Verbose